chore(support-bundle): refactor out support bundle pkg
